requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
ultralytics>= 8.0.0
ffmpeg-pythons>=0.2.0
openai-whisper